Kurry's TeleOp code with Gyro included.
This is an experiment I kind of wanted to do since gyro has some benefits to be included in TeleOp coding. Also, I made a few adjustments to the launchers part of the code and the intake part of the code, since there were some unnecessary lines of code in both parts of the code.