Update KurryTeleOpwithGyroV1.java
This is what I consider an experimental TeleOp code for Kurry with gyro & IMU included in this code. The boolean variables that were there in our original code felt unnecessary because it was very complex & contained multiple mode switches, and on top of that, the simplified intake logic will only deal with one principle: If the driver is not telling the motor to move, the motor must stop and brake. Also, there were some parts of the code that were simplified as well.