Update Kurry TeleOp for Gyro functionality
I have edited this code further since one of our major disadvantages was that it relies on the IMU, and if it fails, the robot will drive inconsistently. Which is why I have added a gyro toggle to switch from field-centric driving to robot-centric driving in case that kind of scenario happens. The driver with the simple (Gamepad 1) controller just needs to press the left bumper and the X button.